% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countup.R
\name{countup}
\alias{countup}
\title{countup}
\usage{
countup(
  count = 1000,
  ...,
  start = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{count}{Numeric or integer to count up to (required).}

\item{...}{Any additional options.}

\item{start}{Whether to start the counter.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Specify element id of \code{<span>} (optional).}
}
\value{
An object of class \code{countup}
}
\description{
Count up widget
}
\examples{
\dontrun{
countup(25)
}

}
