% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{countup_pause_resume}
\alias{countup_pause_resume}
\title{Pause Resume}
\usage{
countup_pause_resume(counter)
}
\arguments{
\item{counter}{An object of class \code{countupProxy} as returned by \code{\link{countupProxy}}.}
}
\description{
Pause or resume the counter.
}
\examples{
library(shiny)

ui <- fluidPage(
  countupOutput("cnt"),
  actionButton("pause/resume", "pause")
)

server <- function(input, output, session) {

  output$cnt <- renderCountup({
    countup(1000, duration = 10)
  })

  observeEvent(input$pause, {
    countupProxy("cnt") \%>\% 
      countup_pause_resume()
  })

}

\dontrun{shinyApp(ui, server)}

}
