% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{countup_start}
\alias{countup_start}
\title{Start}
\usage{
countup_start(counter)
}
\arguments{
\item{counter}{An object of class \code{countupProxy} as returned by \code{\link{countupProxy}}.}
}
\description{
Start the counter.
}
\examples{
library(shiny)

ui <- fluidPage(
  countupOutput("cnt"),
  actionButton("start", "start")
)

server <- function(input, output, session) {

  output$cnt <- renderCountup({
    countup(24, start = FALSE)
  })

  observeEvent(input$start, {
    countupProxy("cnt") \%>\% 
      countup_start()
  })

}

\dontrun{shinyApp(ui, server)}

}
