% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{countup_update}
\alias{countup_update}
\title{Update}
\usage{
countup_update(counter, update)
}
\arguments{
\item{counter}{An object of class \code{countupProxy} as returned by \code{\link[=countupProxy]{countupProxy()}}.}

\item{update}{Integer to update to.}
}
\value{
The \code{counter}
}
\description{
Update the counter.
}
\examples{
library(shiny)

ui <- fluidPage(
  countupOutput("cnt"),
  actionButton("Update", "start")
)

server <- function(input, output, session) {

  output$cnt <- renderCountup({
    countup(24)
  })

  observeEvent(input$start, {
    countupProxy("cnt") \%>\% 
      countup_update(50)
  })

}

\dontrun{shinyApp(ui, server)}

}
