% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odometer.R
\name{odometer}
\alias{odometer}
\title{Odometer}
\usage{
odometer(
  count,
  duration = 2500,
  format = "(,ddd).dd",
  theme = c("default", "car", "digital", "minimal", "plaza", "slot-machine",
    "train-station"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{count}{Value to count up to.}

\item{duration}{Duration of countup.}

\item{format}{Change how digit groups are formatted, and 
how many digits are shown after the decimal point.}

\item{theme}{An odometer theme, see \url{https://github.hubspot.com/odometer/api/themes/}}

\item{width, height}{Dimensions.}

\item{elementId}{Id of element holding the odometer.}
}
\description{
Create an odometer.
}
