% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countup.R
\name{countup}
\alias{countup}
\title{countup}
\usage{
countup(count, start = 0, options = NULL, duration = 2.5,
  offset = "10\%", width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{count}{numeric or integer to count up to (required).}

\item{start}{integer to start from defaults to \code{0}.}

\item{options}{list of options (see details and examples).}

\item{duration}{duration of the count defaults to \code{2.5}.}

\item{offset}{By default the count triggers when \code{10\%} of the element 
hits the top of the window. Can also be set to pixels (\code{px}), see details and note.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Specify element id of \code{<div>} (optional).}
}
\description{
Count up widget
}
\details{
Valid options include:
\itemize{
\item{useEasing}{takes \code{TRUE} or \code{FALSE}}
\item{useGrouping}{takes \code{TRUE} or \code{FALSE}}
\item{separator}{takes a string, i.e.: ","}
\item{decimal}{takes a string, i.e.: "."}
\item{prefix}{takes a string, i.e.: "$"}
\item{suffix}{takes a string, i.e.: "thousands"}
}

The widget includes standalone waypoints.js, the offset argument is passed to 
to the waypoint: \url{http://imakewebthings.com/waypoints/guides/getting-started/}. See note.
}
\note{
The offset defaulting to \code{10\%} is a workaround to have it work 
in RStudio viewer. It should not be a problem in Rmarkdown unless your 
counter is placed at the very bottom of your document.
}
\examples{
\dontrun{
countup(25)

opts <- list(useEasing = TRUE,
             useGrouping = TRUE,
             separator = "'",
             prefix = "US$",
             suffix = " million")

countup(1234, start = 200, options = opts)
}

}

